refactor(perl-dap): split debug adapter patterns into submodule (#0000)
